Fix flaky k-rate-panner test
Change the testing scheme for k-rate panner to verify that the output
is piecewise constant for a constant source when the automation rate
is set to k-rate. Reduce the duration of the test as well; Just a few
render quanta is enough to verify that the output is piece-wise constant
and different from the a-rate output.

Also updated k-rate-constant-source.html to use the same doTest method
since the special case is supported.

Didn't actually figure out the underlying flakiness of the original
test.

Bug: 841856
Change-Id: I8e9807e530fc3b1452bae521a2660be3fbe104f9