Convert word-wrap to an alias of overflow-wrap
word-wrap was previously defined as its own property in
CSSProperties.json5, but with the same behavior as overflow-wrap.
However, this meant that word-wrap showed up alongside word-wrap in
cssText. This patch converts word-wrap to be an alias of overflow-wrap,
as this will prevent the issue of both properties showing up in cssText
and matches the resolution in
https://github.com/w3c/csswg-drafts/issues/866.

Bug: 679068
Change-Id: I3d415eb1cde9a92dffa2a045ba5fa8b015eeb2d7